Move browser logic to BrowserManager
